rename referendum codec file
