+++ .eslintignore
@@ -1,3 +1,3 @@
 dist/
 node_modules/
-src/referendum-codecs.ts
+src/referendum-tx-codecs.ts
